% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AtlasComputeClinicalDataPrediction.R
\name{AtlasComputeClinicalDataPrediction}
\alias{AtlasComputeClinicalDataPrediction}
\title{Build regression model to predict clinical data outcomes from 
node's factors}
\usage{
AtlasComputeClinicalDataPrediction(
  df.clinical.data,
  factors,
  str.clinical.data
)
}
\arguments{
\item{df.clinical.data}{(data frame) clinical data outcomes for each sample.}

\item{factors}{(data frame) factor vector for each node of the map.}

\item{str.clinical.data}{(list) string list of the clinical data to compute 
prediction probabilities.}
}
\value{
predicted probability data frame
}
\description{
We build logistic or linear regression model depending of the
clinical data outcomes nature (binary, categorical or continue).
From this trained model, we predict probability of each map's node to predict
clinical data.
}
