% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AtlasComputeCorrelation.R
\name{AtlasComputeCorrelation}
\alias{AtlasComputeCorrelation}
\title{Compute correlation score between features and clinical data outcomes}
\usage{
AtlasComputeCorrelation(data, clinical_data)
}
\arguments{
\item{data}{(list) list of data frame containing all feature values.}

\item{clinical_data}{(list) string list of the clinical data to compute 
correlation scores.}
}
\value{
list of data frame containing correlation scores
}
\description{
We compute multiple correlation scores between features and
clinical value outcomes. By default, we compute "pearson", "spearman", 
"kendall" and "combined". Note that combined score is a experimental custom
score.
}
