% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AtlasDefineMaxMinClinicalData.R
\name{AtlasDefineMaxMinClinicalData}
\alias{AtlasDefineMaxMinClinicalData}
\title{Define the variable ranges of all features.}
\usage{
AtlasDefineMaxMinClinicalData(data, clinical.data, df, mapdata)
}
\arguments{
\item{data}{(list) list of selected feature to analyse.}

\item{clinical.data}{(string) target clinical data name.}

\item{df}{(data frame) clinical data prediction probability.}

\item{mapdata}{(list) list of data frame containing all feature values.}
}
\value{
data frame with max and min values for each features
}
\description{
Define the variable ranges: maximum and minimum of the feature
and low mean values and high mean values for each features. Low and high mean
values are defined according the clinical data prediction probability (high
if prediction > 0.5 and low if < 0.5).
This values are used to plot radarchart (to define axis scale).
}
