% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AtlasFeatureSelection.R
\name{AtlasFeatureSelection}
\alias{AtlasFeatureSelection}
\title{Select top features based on chosen correlation scores.}
\usage{
AtlasFeatureSelection(
  df.results.correlation,
  clinical.data,
  chosen_methods,
  n = 10
)
}
\arguments{
\item{df.results.correlation}{(data frame) correation scores for each
features and method.}

\item{clinical.data}{(string) target clinical data name.}

\item{chosen_methods}{(string) chosen name of correlation method.}

\item{n}{(int) number of top features to return.}
}
\value{
data frame containing best features
}
\description{
We select n top features based on correlation scores between
all features and target clinical data.
}
