% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AtlasGetClinicalDataPrediction.R
\name{AtlasGetClinicalDataPrediction}
\alias{AtlasGetClinicalDataPrediction}
\title{Load or compute target clinical data prediction probabilities.}
\usage{
AtlasGetClinicalDataPrediction(
  file.path,
  df.codes = NULL,
  list.name.clinical.data = NULL,
  df.clinical.data = NULL
)
}
\arguments{
\item{file.path}{(path) path to load and/or save data frame clinical data
prediction probabilities result.}

\item{df.codes}{(data frame) factor vector for each node of the map.}

\item{list.name.clinical.data}{(list) string list of target clinical data.}

\item{df.clinical.data}{(data frame) if exist, data frame with clinical 
data probabilities prediction already computed.}
}
\value{
data frame with clinical prediction probabilities
}
\description{
Function to check if clinical data are already compute and
saved. If not, compute it and add result to existing data frame result, if
exist or create it. If yes, load and return results.
}
