% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AtlasImmunogram.R
\name{AtlasImmunogram}
\alias{AtlasImmunogram}
\title{High level function to define and plot immunogram based on clinical
data.}
\usage{
AtlasImmunogram(
  df.results.correlation,
  df.clinical.data.prediction.probability,
  clinical.data,
  method,
  n.feature,
  return.immunogram.values = FALSE,
  color_df = NULL,
  banner.start = 1.7,
  banner.thickness = 0.16,
  banner.texte = 2.1,
  plot.title = "Untitled",
  path.plot.save = NULL,
  width = 7,
  height = 7,
  sample.to.plot = NULL,
  plot.max.min = FALSE,
  centerzero = FALSE,
  axistype = 1,
  feature.selection.threshold = NULL,
  feature.selection.limit = NULL,
  labels.circle.coord = 1.2,
  labels.cex = 0.8,
  add.custom.label = TRUE,
  font = 1,
  plot.marge = 0
)
}
\arguments{
\item{df.results.correlation}{(dataframe) data frame with results of all 
correlation between features and target clinical data.}

\item{df.clinical.data.prediction.probability}{(dataframe) data frame with
probability prediction of each map nodes regarding the target clinical data.}

\item{clinical.data}{(str) target clinical data name.}

\item{method}{(str) correlation method to select features.}

\item{n.feature}{(int) number of features for each view to plot in the 
immunogram.}

\item{return.immunogram.values}{(bool) if True, return immunogram values and
don't generate and save the Immunogram.}

\item{color_df}{(dataframe) data frame with the list of views and the defined
colors associated for each view.}

\item{banner.start}{(float) position to start the banner.}

\item{banner.thickness}{(float) thickness of the banner.}

\item{banner.texte}{(float) position of the text around the banner.}

\item{plot.title}{(str) title of the figure.}

\item{path.plot.save}{(str) path to save the figure.}

\item{width}{(float) width of the figure.}

\item{height}{(float) height of the figure.}

\item{sample.to.plot}{(list) list of node names for which to generate the 
immunogram.}

\item{plot.max.min}{(bool) if we generate figure with max and min of the 
datas in the immunogram (default=F).}

\item{centerzero}{(bool) if axis radarchart are centered in zero (default=F).}

\item{axistype}{(int) type of axis.}

\item{feature.selection.threshold}{(float) Threshold to select features. We
keep only features with correlation bigger that this threshold. If NULL, we 
don't use threshold.}

\item{feature.selection.limit}{(int) NUmber of feature to keep for the 
immunogram. We keep the best features correlated in the defined limit. If
NULL, we don't use this limit.}

\item{labels.circle.coord}{(float) position of the circle on which we will 
place the labels  (default=1.2).}

\item{labels.cex}{(float) numeric character expansion factor; multiplied by
("cex") yields the final character size. NULL and NA are equivalent to 1.0
(default=0.8).}

\item{add.custom.label}{(bool) if we add custom labels around the figure
(default=T).}

\item{font}{(int) font type: 1: normal, 2: bold, 3: italic.}

\item{plot.marge}{(float) margin size to use in the figure (default=0).}
}
\description{
Circle.coord explication: c(inner, outer, text): 
 -inner: (float) x and y coordinate to start circle banner.
 -outer: (float) x and y coordinate to end circle banner 
 (difference between inner and outer define the thickness of the banner). 
 -text: (float) text coordinate (x, y) around of the banner.
 #' Filter selected features by treshold or range: - by threshold: 
 choose a value and we keep only features with correlation > threshold. 
 - by range: we choose a range limit (int) and we keep only the n most 
 correlated features all views combined.
}
