% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AtlasResultCorrelation.R
\name{AtlasResultCorrelation}
\alias{AtlasResultCorrelation}
\title{High-level function designed to generate an Immunogram from target
clinical data.}
\usage{
AtlasResultCorrelation(
  file.path,
  list.clinical.data = NULL,
  mapdata = NULL,
  df.clinical.data.prediction.probability = NULL
)
}
\arguments{
\item{file.path}{(string) the csv path of data frame with 
saved clinical data probabilities prediction.}

\item{list.clinical.data}{list of names of clinical data for which 
probabilities are to be calculated. If we}

\item{mapdata}{(rds file) large list with all views and for each view a data
frame with faeture values (float).}

\item{df.clinical.data.prediction.probability}{(data frame)}
}
\description{
This high-level function is designed to generate an Immunogram 
from the desired data.
Many options are available for calculating or retrieving interest feature to
generate the figure according to the chosen target.
Input is target clinical data as data frame and the values of all features
to compute correlation scores.
}
