% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AtlasSOMAllClusteringMatrixAgreement.R
\name{AtlasSOMAllClusteringMatrixAgreement}
\alias{AtlasSOMAllClusteringMatrixAgreement}
\title{Compute matrix agreement of the "nb.SOM.maps" clustering results.}
\usage{
AtlasSOMAllClusteringMatrixAgreement(
  path_save,
  path.maps.model.cluster,
  nb.SOM.maps,
  nb.cluster,
  nb.samples,
  compare.models = F,
  path.models.compare = NULL,
  row.names.remove.constante = T,
  clustering.method = "hierarchical_clustering"
)
}
\arguments{
\item{path_save}{path to save the global clustering matrix.}

\item{path.maps.model.cluster}{path to load existing clusters.csv for each 
generated map.}

\item{nb.SOM.maps}{(int) number of generated map.}

\item{nb.cluster}{(description)int) number of cluster}

\item{nb.samples}{(int) number of sample}

\item{compare.models}{(bool) if we want to compare current clustering with
clustering of other model for same sample.}

\item{path.models.compare}{path to the clustering matrix of the model to 
compare.}

\item{row.names.remove.constante}{(bool) if we want clean row names by 
removing all constant nomenclature.}

\item{clustering.method}{(description)str) clustering method name 
(default=hierarchical_clustering)}
}
\value{
global.clustering.matrix
}
\description{
We compute all clustering matrix with shape (n_sample, 
n_seed*n_cluster).
For each clustering loop (among the nb.SOM.maps made), we provide clustering
result in this big matrix. Column names is loop number + cluster number.
This is a binary matrix with 1 for cluster belonging and zero other.
}
