% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AtlasSOMAllMapsWorkflow.R
\name{AtlasSOMAllMapsWorkflow}
\alias{AtlasSOMAllMapsWorkflow}
\title{High level function to create all map.rds for model analysis}
\usage{
AtlasSOMAllMapsWorkflow(
  nb.factor,
  path.maps,
  path.model.factor,
  model.name,
  nb.SOM.maps = 1000,
  first.number.map = 1,
  parallel.run = NULL,
  regularized = F,
  n.grid = 20,
  std = 0.5,
  list.nb.clusters = NULL,
  outliers.substitution = T,
  factor.data.augmentation = 20,
  decision.help.nb.cluster = F,
  super.som = F,
  clustering.method = "hierarchical_clustering",
  first.nb.cluster.SOM.generated = 3
)
}
\arguments{
\item{nb.factor}{(int) number of factor for the current MOFA}

\item{path.maps}{path to maps folder}

\item{path.model.factor}{path to the model directory}

\item{model.name}{name of the current model}

\item{nb.SOM.maps}{number of maps to generate}

\item{first.number.map}{(int) default=1, if the first seed to test}

\item{parallel.run}{config to run script in parallel}

\item{regularized}{(bool) regularized data or not}

\item{n.grid}{(int) size of SOM map}

\item{std}{(float) standard deviation}

\item{list.nb.clusters}{(list) list of all cluster to test}

\item{outliers.substitution}{(bool) to reduce outliers or not}

\item{factor.data.augmentation}{(int) factor augmentation for data 
augmentation}

\item{decision.help.nb.cluster}{(bool) if we plot figure for help decision 
about best number of cluster to choose.}
}
\description{
High level function to create all map.rds and pdfs (codes map, 
clusters map). We do that for the current model passed in argument.
}
