% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AtlasSOMARIMapSelection.R
\name{AtlasSOMARIMapSelection}
\alias{AtlasSOMARIMapSelection}
\title{We use adjusted rand index score to find most similar clustering
of the clustering consensus.}
\usage{
AtlasSOMARIMapSelection(
  path.maps.baseline.cluster,
  path.consensus.clustering,
  clustering.consensus,
  nb.SOM.maps,
  nb.maps.to.keep = 50
)
}
\arguments{
\item{path.maps.baseline.cluster}{path to cluster folder.}

\item{path.consensus.clustering}{path to folder with consensus information.}

\item{clustering.consensus}{data frame with clustering consensus.}

\item{nb.SOM.maps}{number of maps to generate.}

\item{nb.maps.to.keep}{(int) number of best maps to keep.}
}
\description{
We save a data frame with a list of the n best maps contains 
clustering the most similar as consensus clustering. We measure the similarity
with adjusted rand index score.
}
