% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AtlasSOMBestScoreMap.R
\name{AtlasSOMBestScoreMap}
\alias{AtlasSOMBestScoreMap}
\title{Compute three scores to compare all maps clustering}
\usage{
AtlasSOMBestScoreMap(
  nb.cluster,
  nb.SOM.maps,
  path.maps.baseline,
  path.maps.consensus,
  nb.clusters,
  clustering.consensus.outputs,
  model.name,
  compare.models = F,
  ARI.score = T,
  matrix.trace.score = T,
  nb.maps.to.keep = 50,
  jaccard.index = F
)
}
\arguments{
\item{nb.cluster}{(int) number of cluster to test.}

\item{nb.SOM.maps}{number of maps to generate.}

\item{path.maps.baseline}{path to cluster folder.}

\item{path.maps.consensus}{path to save csv.}

\item{nb.clusters}{(int) number of cluster.}

\item{clustering.consensus.outputs}{data frame with clustering consensus.}

\item{model.name}{name of current model.}

\item{compare.models}{(bool) if we compare two models factors.}

\item{ARI.score}{(bool) if we use Adjust Rand Index score to find best map.}

\item{matrix.trace.score}{(bool) if we use matrix trace score to find best 
map.}

\item{nb.maps.to.keep}{(int) number of best maps to keep.}

\item{jaccard.index}{(bool) if use jaccard index (default=F).}
}
\description{
Build matrix of shape (nb.SOM.maps, 3) with scores for each map.
}
