% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AtlasSOMClusterMatrixToSpread.R
\name{AtlasSOMClusterMatrixToSpread}
\alias{AtlasSOMClusterMatrixToSpread}
\title{read and build cluster matrix for current nth.map}
\usage{
AtlasSOMClusterMatrixToSpread(
  path.maps.model.cluster.nth.map,
  delim = ";",
  escape_double = FALSE,
  trim_ws = TRUE,
  matrix.agreement = F,
  nth.map = NULL,
  nb.cluster = NULL
)
}
\arguments{
\item{path.maps.model.cluster.nth.map}{path to maps folder}

\item{delim}{character delimiter for read file}

\item{escape_double}{(bool) default: FALSE}

\item{trim_ws}{(bool) default: TRUE}

\item{matrix.agreement}{(bool) Complete documentation argument}

\item{nth.map}{current map number}

\item{nb.cluster}{(int) number of cluster for data frame shape.}
}
\value{
clusters
}
\description{
Transform clusters.csv data frame in a matrix with nb_cluster
columns (column name: "map_1_cl").
Each cell contains the number of the node if the associated sample is part
of the cluster of the column and contains 0 otherwise.
}
