% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AtlasSOMCompareModels.R
\name{AtlasSOMCompareModels}
\alias{AtlasSOMCompareModels}
\title{Compare clustering of two models for same samples}
\usage{
AtlasSOMCompareModels(data, path.other.model.clustering)
}
\arguments{
\item{data}{current data frame with all clusters groups}

\item{path.other.model.clustering}{path to the csv file of clustering for 
the model to compare}
}
\value{
data
}
\description{
This function compare clustering of two models.
We add in sample name the cluster number of other model for the same sample.
We fill in the cluster group number of the other model by add 
("_cl%d", cluster_group) in the sample name.
}
