% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AtlasSOMComputeClusteringConsensus.R
\name{AtlasSOMComputeClusteringConsensus}
\alias{AtlasSOMComputeClusteringConsensus}
\title{Compute consensus clustering from jaccard distances}
\usage{
AtlasSOMComputeClusteringConsensus(
  jaccard_distances_all_clusters,
  path_save,
  n.cluster,
  nb.SOM.maps,
  clustering.method = "hierarchical_clustering",
  plot_title_jaccard_distances = "heatmap_global_jaccard_sample"
)
}
\arguments{
\item{path_save}{path to save pdf and csv}

\item{nb.SOM.maps}{(int) number of generated map}

\item{plot_title_jaccard_distances}{title of the plot Jaccard distances}

\item{global_clustering_matrix}{data frame with all clustering results}

\item{nb_cluster}{current number of cluster}
}
\value{
clustering_consensus, jaccard_distances_all_clusters
}
\description{
This function compute a dataframe with the consensus clustering
based on all clustering made and jaccard distances between them.
First, we compute the Jaccard distances to measure the dissimilarity between
all clusters in global cluster matrix.
Then, we cut the obtained dendrogramme with nb_cluster variable.
}
