% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AtlasSOMConsensusMap.R
\name{AtlasSOMConsensusMap}
\alias{AtlasSOMConsensusMap}
\title{High level function to analyze created maps.}
\usage{
AtlasSOMConsensusMap(
  nb.cluster,
  path.maps.baseline,
  path.maps.consensus,
  nb.SOM.maps,
  nb.samples,
  model.name.factor,
  clustering.method = "hierarchical_clustering",
  compare.models = F,
  path.models.compare = NULL,
  row.names.remove.constante = T
)
}
\arguments{
\item{nb.cluster}{(int) number of cluster to test.}

\item{path.maps.baseline}{path to baseline maps folder}

\item{path.maps.consensus}{path to consensus maps folder}

\item{nb.SOM.maps}{number of maps to generate}

\item{model.name.factor}{model name with number of factors.}

\item{clustering.method}{(str) clustering method used.}

\item{compare.models}{(bool) if we compare two models factors.}

\item{path.models.compare}{path to the models to compare.}

\item{row.names.remove.constante}{(bool) if we want clean row names by 
removing all constant nomenclature.}
}
\description{
High level function to analyse all created maps
}
