% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AtlasSOMCreateGroupCluster.R
\name{AtlasSOMCreateGroupCluster}
\alias{AtlasSOMCreateGroupCluster}
\title{Compute consensus clustering}
\usage{
AtlasSOMCreateGroupCluster(
  data.som,
  path.map.cluster.nth.map.clustering,
  nb.cluster = NULL,
  decision.help.nb.cluster = F,
  clustering.method = "hierarchical_clustering"
)
}
\arguments{
\item{data.som}{(.rds) SOM map}

\item{path.map.cluster.nth.map.clustering}{path to the cluster.csv file for
the nth map.}

\item{nb.cluster}{(int) number of cluster groups.}

\item{decision.help.nb.cluster}{(bool) if we plot figure for help decision 
about best number of cluster to choose.}

\item{clustering.method}{(str) clustering method used.}
}
\value{
list of group for each sample
}
\description{
This function determine each cluster group for samples.
First, this function compute the euclidean distance matrix between codes's
SOM map.
Code is the embedding of samples (using MOFA) and each code are compute for
all nodes of the map.
Then, Hierarchical clustering analysis on this distances is made with the
ward.2 algorithm.
If the number of cluster is not provided, we compute the best number of
cluster based on the hierarchical clustering.
Then, we cut the obtained dendrogramme with nb.cluster variable.
This function can also compute kmeans clustering with ConsensusClusterPlus
algorithm.
}
