% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AtlasSOMDataOutliersSubstitution.R
\name{AtlasSOMDataOutliersSubstitution}
\alias{AtlasSOMDataOutliersSubstitution}
\title{Replaces outliers values in the data frame.}
\usage{
AtlasSOMDataOutliersSubstitution(data, low = NULL, up = NULL, by_column = TRUE)
}
\arguments{
\item{data}{data to be modified.}

\item{low}{(float) value of lower boundary (the values below "low" will be 
replaced by "low").}

\item{up}{(float) value of upper boundary (the values above "up" will be 
replaced by up").}

\item{by_column}{(boolean) Replace outliers independently by column 
(recommended).}
}
\value{
data frame modified
}
\description{
Function to replace outliers.
By default we compute Low and Up values from quantiles (0.25 and 0.75).
Then the values above and below boundaries are replaced by these values.
Default boundaries values are:
    Low:  Q3 + 3*IQR
    Up: Q - 3*IQR
}
