% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AtlasSomKmeansConsensus.R
\name{AtlasSomKmeansConsensus}
\alias{AtlasSomKmeansConsensus}
\title{Function to compute Kmeans clustering with ConsensusClusterPlus 
algorithm.}
\usage{
AtlasSomKmeansConsensus(
  path.result,
  distances,
  nb.cluster.min = 2,
  nb.cluster.max = 10,
  clusterAlg = "km",
  distance = "euclidean",
  reps = 100,
  pItem = 0.8,
  return.nb.cluster = 0
)
}
\arguments{
\item{path.result}{(str) path to save result.}

\item{distances}{(list) data to be clustered.}

\item{nb.cluster.min}{(int) minimum number of cluster (default=2).}

\item{nb.cluster.max}{(int) maximum number of cluster (default=10).}

\item{clusterAlg}{(str) name of the algorithm to use (default=km).}

\item{distance}{(str) type of distance to use (default=euclidean).}

\item{reps}{(int) iteration number for ConsensusClusterPlus algortihm.}

\item{pItem}{see ConsensusClusterPlus documentation}

\item{return.nb.cluster}{(int) condition to return result only if we have at
least 1 cluster.}
}
\description{
This function use ConsensusClusterPlus algorithm to compute the
clustering consensus of distances SOM nodes.
If we don't provide number of cluster, we run the algorithm for each number
of cluster of the list cluster number.
Then, outputs are save in Map folder path.
Then, it need human supervisor to do number of cluster choice.
}
