% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AtlasSOMOptimalClusterNumber.R
\name{AtlasSOMOptimalClusterNumber}
\alias{AtlasSOMOptimalClusterNumber}
\title{Compute the best number of cluster.}
\usage{
AtlasSOMOptimalClusterNumber(
  return.best.nb.cluster = T,
  data.som.codes = NULL,
  hierrarchical.cluster.analysis = NULL,
  path = NULL,
  min.nb.cluster = 2,
  max.nb.cluster = 20,
  plot.graph = T,
  return.losses.list = FALSE,
  plot.Elbow = T,
  plot.silhouette = T
)
}
\arguments{
\item{return.best.nb.cluster}{(bool) if we want return the value of optimal 
cluster number computes with our algorithm.}

\item{data.som.codes}{matrix with factors codes of SOM map.
#' @param hierrarchical.cluster.analysis hierarchical clustering of data as 
data frame.}

\item{path}{path to save plot.}

\item{min.nb.cluster}{(int): minimum of cluster possible}

\item{max.nb.cluster}{(int): maximum of cluster possible}

\item{plot.graph}{(bool): to plot it or not}

\item{return.losses.list}{(bool): if return losses list or not}

\item{plot.Elbow}{(bool): if compute and plot Elbow scores to decision help 
regarding best number of cluster}

\item{plot.silhouette}{(bool): if compute and plot Silhouette scores to*
decision help regarding best number of cluster}
}
\value{
either the list of losses and the optimal number of cluster (k) or 
only k.
}
\description{
Algorithm to find out the best k number of cluster to cut
hierarchical clustering tree. We wan also generate plot for decision help
with Elbow and silhouette scores.
}
