% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AtlasSOMPrimaryAnalysis.R
\name{AtlasSOMPrimaryAnalysis}
\alias{AtlasSOMPrimaryAnalysis}
\title{Generate Primary SOM map analysis and save it as figure.}
\usage{
AtlasSOMPrimaryAnalysis(
  data.som,
  nb.samples,
  nb.cluster,
  path.map.cluster.nth.map,
  sample.groups = NULL,
  hierarchical.clustering = F,
  kmeans.clustering = T
)
}
\arguments{
\item{data.som}{(rds format) SOM map.}

\item{nb.samples}{(int) number of sample data/}

\item{path.map.cluster.nth.map}{path to the nth SOM map for specific MOFA
model.}

\item{sample.groups}{cluster sample group}

\item{hierarchical.clustering}{(bool) if we use hierarchical clustering 
algorithm}

\item{kmeans.clustering}{(bool) if we use kmeans clustering algorithm}

\item{nb.cluster(int)}{number of cluster}
}
\description{
We build and save basics SOM map analysis as figure from the 
passed in argument MOFA model or we load it by path passed in argument.
}
