% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AtlasSOMSensibility.R
\name{AtlasSOMSensibility}
\alias{AtlasSOMSensibility}
\title{Compute the sensibility of cluster_consensus.}
\usage{
AtlasSOMSensibility(cluster_consensus, jaccard_distances_all_clusters)
}
\arguments{
\item{cluster_consensus}{data frame with clustering consensus results}

\item{jaccard_distances_all_clusters}{jaccard distances between all clusters}
}
\value{
mean sensibility
}
\description{
Compute the sensibility of cluster_consensus.
Distance between these samples.
This sum should be as small as possible because we want the distance to be 
as small as possible for samples of the same cluster.
}
