% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AtlasSOMSpecificty.R
\name{AtlasSOMSpecificty}
\alias{AtlasSOMSpecificty}
\title{Compute the specificity of cluster_consensus.}
\usage{
AtlasSOMSpecificty(cluster_consensus, jaccard_distances_all_clusters)
}
\arguments{
\item{cluster_consensus}{data frame with clustering consensus results}

\item{jaccard_distances_all_clusters}{jaccard distances between all clusters}
}
\value{
mean specificity
}
\description{
Compute the specificity of cluster_consensus.
Here we add the distances of clusters that are not clustered together.
We want this value to be as high as possible because two samples that are not 
from the same cluster should have a large distance.
}
