% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AtlasSOMUnitMapCreation.R
\name{AtlasSOMUnitMapCreation}
\alias{AtlasSOMUnitMapCreation}
\title{Create and save SOM map as rds format.}
\usage{
AtlasSOMUnitMapCreation(
  data,
  n.grid = 20,
  radius = NULL,
  topo = "hexagonal",
  toroidal = F,
  maxNA.fraction = 1,
  rlen = 2000,
  alpha = c(0.5, 0.01),
  super.som = F
)
}
\arguments{
\item{data}{matrix or data frame containing the input data to be used in the 
Self-Organizing Map (SOM) algorithm.}

\item{n.grid}{(int) number of nodes or neurons to be used in the 
Self-Organizing Map (SOM) algorithm.}

\item{radius}{neighborhood radius used in the Self-Organizing Map (SOM)
algorithm. The neighborhood radius determines the size of the neighborhood 
around the winning neuron or node that will be updated during each training 
iteration.
During the training process, the neighborhood radius is gradually decreased 
over time as the SOM converges to a stable configuration. This allows the 
SOM to start with a large neighborhood radius that covers the entire map,
and then gradually narrow the focus to smaller areas as the training 
progresses.}

\item{topo}{type of neighborhood function to use in the Self-Organizing Map
(SOM) algorithm.}

\item{toroidal}{(bool) specifies whether the Self-Organizing Map (SOM) grid
should be toroidal or not}

\item{maxNA.fraction}{maximum fraction of missing values allowed in the input
data. If the maxNA.fraction argument is set to a value between 0 and 1, 
the SOM algorithm will automatically remove any columns (variables) from the 
input data that have a fraction of missing values greater than maxNA.fraction.}

\item{rlen}{(int) specifies the number of iterations or epochs that the 
Self-Organizing Map (SOM) algorithm will be run for during training.}

\item{alpha}{specifies the learning rate or step size used during the 
training process of the Self-Organizing Map (SOM) algorithm.}
}
\value{
data.som generated
}
\description{
Create Kohonen SOM map with argument parameters and save it.
}
