% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AtlasSOMUnitMapWorkflow.R
\name{AtlasSOMUnitMapWorkflow}
\alias{AtlasSOMUnitMapWorkflow}
\title{Create and save SOM map as rds format and plot to analyse map.}
\usage{
AtlasSOMUnitMapWorkflow(
  nth.map,
  data,
  path.map.cluster,
  path.first.cluster,
  nb.cluster = NULL,
  n.grid = 20,
  regularized = FALSE,
  std = 0.5,
  factor.data.augmentation = 20,
  first.cluster = F,
  decision.help.nb.cluster = F,
  super.som = F,
  return.clusters = F,
  save.outputs = T,
  map.analysis = T,
  clustering.method = "hierarchical_clustering"
)
}
\arguments{
\item{nth.map}{(int) current nth.map.}

\item{data}{rds file.}

\item{path.map.cluster}{path to save map for specific cluster number.}

\item{path.first.cluster}{path where maps with the first number of cluster 
to test was saved.}

\item{n.grid}{(int) size of SOM map.}

\item{regularized}{(bool) regularized data or not.}

\item{std}{(float) standard deviation.}

\item{factor.data.augmentation}{(int) factor augmentation for data
augmentation.}

\item{first.cluster}{the first number of cluster to test.}

\item{decision.help.nb.cluster}{(bool) if we plot figure for help decision 
about best number of cluster to choose.}

\item{super.som}{(bool) if we use supersom instead som algorithm (default=F).}

\item{return.clusters}{(bool) if function return clusters result (default=F).}

\item{save.outputs}{(bool) if save the result (default=F).}

\item{map.analysis}{(bool) if we make analysis map (default=T).}

\item{clustering.method}{(str) name of clustering method to use
(default=hierarchical_clustering).}

\item{nb.cluster(int)}{number of cluster.}
}
\description{
Create Kohonen's self-organising maps from Kohonen R package.
This map is saved as .rds file in the define "som_save_path" destination.
Data in argument is usually model's Zmatrix with shape (n_sample, n_factors).
By default, parameters of SOM are already defined, but it is possible to
entire defined it by passed this arguments as dictionary.
Artificial data augmentation based on Gaussian noise is made to perform
Kohonen SOM algorithm. The data augmentation multiplication factor is
"factor.data.augmentation".
}
