% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CustomCheckFileExistence.R
\name{CustomCheckFileExistence}
\alias{CustomCheckFileExistence}
\title{Check if file exist with custom Atlas assistance}
\usage{
CustomCheckFileExistence(path_to_check, create_dir_if_missing = T)
}
\arguments{
\item{path_to_check}{path to the file to check existence.}

\item{create_dir_if_missing}{(bool) if we want to create directory if missing}
}
\value{
(bool) T or F depending file or directory existence
}
\description{
This function check if file exist and depending of request,
custom help is provide.
If we check directory existence, by default if is missing we create it.
e.g.: if we check existence of specific file and this file is missing,
help assistance will provide the function name to be run to generate missing
file.
}
