% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataFormatCheckMOFA.R
\name{DataFormatCheckMOFA}
\alias{DataFormatCheckMOFA}
\title{Data format check for MOFA algorithm.}
\usage{
DataFormatCheckMOFA(data, epsilon = 1e-04)
}
\arguments{
\item{data}{(data frame) data to check}

\item{epsilon}{(float) by default: 1e-4, is the epsilon.}
}
\value{
(bool) True if all tests passed.
}
\description{
We test data have good format to be used in MOFA
algorithm. 
The following tests are performed: 
1. we test that the datarame has the following 5 columns: “sample”, 
“feature”, “view”, “value”, “group”.
2. we test that all "feature name + view name" are unique.
3. We test that the values across views are centered reduce 
(with a epsilon of 1e-4).
4. we test that the feature name across a view are not duplicated.
5. We test that they have at least 2 values per views.
}
