% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetSeed.R
\name{.SetSeed}
\alias{.SetSeed}
\title{Private function to set seed for reproductibility}
\usage{
.SetSeed(seed = 42)
}
\arguments{
\item{seed}{(int): by default is 42}
}
\description{
Function to define same seed for all project.
}
