% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitializationFolderProject.R
\name{InitializationFolderProject}
\alias{InitializationFolderProject}
\title{Create project architecture folder for atlastoolsv2 project.}
\usage{
InitializationFolderProject()
}
\description{
This function create multiple folder to run Atlas projects and
save outputs.
The following folders are created (if not exist):
data, data/external, data/processed, data/raw.
models.
reports, reports/scripts_outputs, reports/figures, reports/seed,
reports/maps, reports/maps/1_Baseline_analysis, 
reports/maps/2_Consensus_analysis, reports/maps/3_Best_maps,
reports/maps/4_Best_maps_consensus, reports/maps/5_Selected_maps.
}
