% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitializeParallelRun.R
\name{InitializeParallelRun}
\alias{InitializeParallelRun}
\title{Set up parallel settings for computation.}
\usage{
InitializeParallelRun(n = 0.8, max = F)
}
\arguments{
\item{n}{(float): percent of available cores}

\item{max}{(bool) if we use th emaximum ressources allowed.}
}
\value{
makeCluster(nb_cores_to_use)
}
\description{
Function to automatically allocated n percent of available
cores to parallel running.
By default we use 80% of available cores. If max=T, we use all available
cores except one.
}
