% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOFAModelAnalysis.R
\name{MOFAModelAnalysis}
\alias{MOFAModelAnalysis}
\title{Several MOFA model analysis to explain model capacity.}
\usage{
MOFAModelAnalysis(
  model,
  path.result,
  Clinical = NA,
  perform.GSEA = F,
  not.denoise = NULL,
  list.view.GSEA = NULL
)
}
\arguments{
\item{model}{(MOFA) MOFA model trained containing weights to extract.}

\item{path.result}{path to save generated plot}

\item{Clinical}{(bool) to generate additional clinical-factor correlation}

\item{perform.GSEA}{(bool) to generate additional plot about GSEA}

\item{not.denoise}{(list) list of view that should not be denoised.}

\item{list.view.GSEA}{(list) we generate GSEA outputs only for feature in
this list.}
}
\description{
We explore different measure of model capacity to understand
how work this model and his performances.
}
