% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOFAModelCollapsePathway.R
\name{MOFAModelCollapsePathway}
\alias{MOFAModelCollapsePathway}
\title{Function title to complete}
\usage{
MOFAModelCollapsePathway(
  fgseaRes,
  pathways,
  stats,
  pval.threshold = 0.05,
  nperm = 10/pval.threshold,
  gseaParam = 1
)
}
\arguments{
\item{nperm}{description}

\item{gseaParam}{description}
}
\value{
list
}
\description{
Function description to complete
}
