% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOFAModelExtraAlnalysis.R
\name{MOFAModelExtraAlnalysis}
\alias{MOFAModelExtraAlnalysis}
\title{additional plot to compare different phases model}
\usage{
MOFAModelExtraAlnalysis(
  list.factor,
  path.model,
  model_phase1_name,
  model_phase2_name
)
}
\arguments{
\item{list.factor}{(lsit) list of factor number tested for MOFA mdoel.}

\item{path.model}{path to save generated plot}

\item{model_phase1_name}{path to model for the first phase}

\item{model_phase2_name}{path to model for the second phase}
}
\description{
we compare several factors models in different phase.
}
