% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOFAModelExtractCumulativeR2.R
\name{MOFAModelExtractCumulativeR2}
\alias{MOFAModelExtractCumulativeR2}
\title{This function extract explained variance of MOFA model.}
\usage{
MOFAModelExtractCumulativeR2(model)
}
\arguments{
\item{model}{MOFA model to extract variance.}
}
\description{
This function extract explained variance of the MOFA model 
and compute the cumulative variance through factors.
}
