% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOFAModelExtractR2.R
\name{MOFAModelExtractR2}
\alias{MOFAModelExtractR2}
\title{Extract R2 explained variance from MOFA model.}
\usage{
MOFAModelExtractR2(path.model)
}
\value{
result.additional (data frame) additional explained variance by 
factor.
}
\description{
We extract R2 explained variance from all MOFA models with 
different number of factor. Then, we plot additional explained variance by 
factor.
}
