% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOFAModelFactorsAnalysis.R
\name{MOFAModelFactorsAnalysis}
\alias{MOFAModelFactorsAnalysis}
\title{Explore top factor correlated to features and plot figures.}
\usage{
MOFAModelFactorsAnalysis(
  path.model,
  path.model.factor,
  keep.group = NULL,
  show_colnames = T
)
}
\arguments{
\item{path.model}{path to the trained model to load}

\item{path.model.factor}{path to the factor folder of the current model}

\item{keep.group}{(bool) If we want keep only samples of specific group}

\item{show_colnames}{(bool) If TRUE show sample names on heatmaps}
}
\description{
We compute top features correlated to the different factors.
Then we analyse this correlation and plot figures.
}
