% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOFAModelGSEACalcul.R
\name{MOFAModelGSEACalcul}
\alias{MOFAModelGSEACalcul}
\title{Function title to complete}
\usage{
MOFAModelGSEACalcul(gene_list, myGO, pval, name = NULL)
}
\value{
output
}
\description{
Function description to complete
}
