% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOFAModelSeedSearch.R
\name{MOFAModelSeedSearch}
\alias{MOFAModelSeedSearch}
\title{Train n models and keep the best one according the ELBO score.}
\usage{
MOFAModelSeedSearch(
  nb.seed,
  path.model.factor,
  data = NULL,
  drop.factor = NULL,
  factor.number = 20,
  max.iteration = 15000,
  keep.group = NULL,
  scale.view = F,
  scale.group = F
)
}
\arguments{
\item{nb.seed}{(int) Number of different seed to test in order to find out 
the best MOFA model training initialization.}

\item{path.model.factor}{(file.path) path to save model with specific factor}

\item{data}{data frame on which to train the MOFA model}

\item{drop.factor}{(float) numeric indicating the threshold on fraction of
variance explained to consider a factor inactive and drop it from the model.
For example, a value of 0.01 implies that factors explaining less than 1% of
variance (in each view) will be dropped. Default is -1 (no dropping of 
factors)}

\item{factor.number}{(int) current number of factors for MOFA model}

\item{max.iteration}{(int) numeric value indicating the maximum number of
iterations. Default in this function is 15000. 
Convergence is assessed using the ELBO statistic}

\item{keep.group}{(list) custom parameter to train MOFA model only on the 
subset of data belonging to the groups in the list. 
By default it is "NULL", which means that the MOFA model is trained on
the whole data set (all groups).}

\item{scale.view}{(bool) if views have different ranges/variances, it is 
good practice to scale each view to unit variance. Default is TRUE here.}

\item{scale.group}{(bool)  if groups have different ranges/variances, it is 
good practice to scale each group to unit variance. Default is TRUE here.}
}
\description{
We Train n MOFA models with different initialization seed.
Then we compare ELBO score between different MOFA model and we keep
the best one regarding the ELBO score.
}
