% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOFAModelsFolderManager.R
\name{MOFAModelsFolderManager}
\alias{MOFAModelsFolderManager}
\title{Manage creation and/or deletion of models directories to respect 
folder structure.}
\usage{
MOFAModelsFolderManager(
  path.model,
  path.model.factor,
  Clear.models_results.existing,
  model.name
)
}
\arguments{
\item{path.model}{(str) path to the folder of model to manage.}

\item{path.model.factor}{(str) path to the factor folder of current model}

\item{Clear.models_results.existing}{(bool) if True,  we delete every files
and object and directories in existing folder. If False, we copy all 
existing files, objects and directory in folder called "old_model".}

\item{model.name}{(str) name of the current model.}
}
\description{
But by default, we copy last existing models and outputs model analysis in 
folder called "old_model".
This function, can delete whole existing directories if 
'Clear.models_results.existing' are True.
}
