% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOFAModelTraining.R
\name{MOFAModelTraining}
\alias{MOFAModelTraining}
\title{Train MOFA models with MOFA option.}
\usage{
MOFAModelTraining(
  data,
  path,
  drop.factor = 0.001,
  mode.convergence = "slow",
  max.iteration = 15000,
  factor.number = 13,
  scale.view = F,
  scale.group = F,
  keep.group = NULL,
  seed = 1
)
}
\arguments{
\item{data}{(data frame) multi-omics data set to train MOFA model}

\item{path}{(file.path) MOFA models and outputs directory path}

\item{drop.factor}{(float) numeric indicating the threshold on fraction of
variance explained to consider a factor inactive and drop it from the model.
For example, a value of 0.01 implies that factors explaining less than 1% of
variance (in each view) will be dropped. Default is -1 (no dropping of 
factors)}

\item{mode.convergence}{mode of convergence.}

\item{max.iteration}{(int) numeric value indicating the maximum number of
iterations. Default in this function is 15000. 
Convergence is assessed using the ELBO statistic}

\item{factor.number}{(int) current number of factors for MOFA model}

\item{scale.view}{(bool) if views have different ranges/variances, it is 
good practice to scale each view to unit variance. Default is TRUE here.}

\item{scale.group}{(bool)  if groups have different ranges/variances, it is 
good practice to scale each group to unit variance. Default is TRUE here.}

\item{keep.group}{(list) custom parameter to train MOFA model only on the 
subset of data belonging to the groups in the list. 
By default it is "NULL", which means that the MOFA model is trained on
the whole data set (all groups).}

\item{seed}{(int) current seed value for MOFA training initialization}
}
\description{
We define all data, model, training options and we train MOFA
models.
}
