% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOFAModelVarianceAnalysis.R
\name{MOFAModelVarianceAnalysis}
\alias{MOFAModelVarianceAnalysis}
\title{Compute and generate plot about model variance}
\usage{
MOFAModelVarianceAnalysis(path.model, results = NULL)
}
\arguments{
\item{path.model}{directory to save plot}
}
\description{
Compute R2 variance and cumulative explained variance and plot
several pdf about variance.
}
