% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOFAModelZmatrixExport.R
\name{MOFAModelZmatrixExport}
\alias{MOFAModelZmatrixExport}
\title{Extract and save Z matrix of MOFA model}
\usage{
MOFAModelZmatrixExport(model, path.save, regularized = F)
}
\arguments{
\item{model}{(MOFA) MOFA model trained containing weights to extract.}

\item{path.save}{path to save the Z matrix}

\item{regularized}{(bool) if we regularized weights of the weights matrix}
}
\description{
Load MOFA model trained and extract Z matrix as data frame, then
this Z matrix is saved as .RDS file.
}
