% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOFAMultiTraining.R
\name{MOFAMultiTraining}
\alias{MOFAMultiTraining}
\title{Train MOFA model with different seed to find the best one.}
\usage{
MOFAMultiTraining(
  nb.factor,
  data,
  path.model = "models/unnamed_model",
  model.name = NULL,
  drop.factor = 0.001,
  keep.group = NULL,
  max.iteration = 15000,
  MOFA.nb.seed = 20,
  scale.view = F,
  scale.group = F,
  Clear.models_results.existing = F,
  not.denoise = NULL,
  show_colnames = T
)
}
\arguments{
\item{nb.factor}{(int) current number of factor for MOFA model to train.}

\item{data}{(data frame) multi-omics data set to train MOFA model}

\item{path.model}{(file.path) path to save trained model and analysis model}

\item{model.name}{(str) name of the current model}

\item{drop.factor}{(float) numeric indicating the threshold on fraction of
variance explained to consider a factor inactive and drop it from the model.
For example, a value of 0.01 implies that factors explaining less than 1% of
variance (in each view) will be dropped. Default is -1 (no dropping of 
factors)}

\item{keep.group}{(list) custom parameter to train MOFA model only on the 
subset of data belonging to the groups in the list. 
By default it is "NULL", which means that the MOFA model is trained on
the whole data set (all groups).}

\item{max.iteration}{(int) numeric value indicating the maximum number of
iterations. Default in this function is 15000. 
Convergence is assessed using the ELBO statistic.}

\item{MOFA.nb.seed}{(int) Number of different seed to test in order to find 
out the best MOFA model training initialization. By default (here) is 20.}

\item{scale.view}{(bool) if views have different ranges/variances, it is 
good practice to scale each view to unit variance. Default is TRUE here.}

\item{scale.group}{(bool) if groups have different ranges/variances, it is 
good practice to scale each group to unit variance. Default is TRUE here.}

\item{Clear.models_results.existing}{(bool) if True,  we delete every files
and object and directories in existing folder. If False, we copy all 
existing files, objects and directory in folder called "old_model".}

\item{not.denoise}{(list) list of view that should not be denoised.}

\item{show_colnames}{(bool) if TRUE show sample names on heatmaps.}
}
\description{
We use different initialization seed to train the best MOFA 
model with multi-omics dataframe.
We use and test different parameters values to find out the best model 
according our dataframe.
The best MOFA model is selected according ELBO score.
}
