% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProjectionAllFeatures.R
\name{ProjectionAllFeatures}
\alias{ProjectionAllFeatures}
\title{projection of features on existing SOM maps}
\usage{
ProjectionAllFeatures(
  path.directory,
  data.som,
  model,
  nb.cluster,
  keep.groups = NULL,
  groups = NULL,
  regularized = T,
  show.samples = F,
  all.features = T,
  add.intercept = T,
  hierarchical.clustering = F,
  kmeans.clustering = T
)
}
\arguments{
\item{path.directory}{path to current directory where save outputs}

\item{data.som}{SOM map codes matrix.}

\item{model}{MOFA model trained}

\item{nb.cluster}{(int) number of cluster}

\item{keep.groups}{specify which views should be kept in the model.
This parameter takes a vector of integers representing the indices of the 
groups that should be kept.
By default, if keep_group = NULL, all groups are kept in the model.}

\item{groups}{specify which groups should be kept in the model.}

\item{regularized}{(bool) if we regularize or not data}

\item{show.samples}{(bool) if show sample on generated plot}

\item{all.features}{(bool) if we do analysis for all features}

\item{add.intercept}{(bool) if we add intercept}

\item{hierarchical.clustering}{(bool) if use hierarchical clustering 
algorithm (default=F)}

\item{kmeans.clustering}{(bool) if use hierarchical clustering 
algorithm (default=T)}
}
\description{
We project feature values on existing map.
}
