% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProjectionClinicalData.R
\name{ProjectionClinicalData}
\alias{ProjectionClinicalData}
\title{Project Clinical data passed in argument in SOM map.}
\usage{
ProjectionClinicalData(
  nb.cluster,
  dt.Atlas,
  dt.Atlas.clinical,
  list.clinical.projection,
  path.maps,
  model.name,
  nb.SOM.maps,
  nb.samples
)
}
\arguments{
\item{nb.cluster}{(int) number of cluster of the current map.}

\item{dt.Atlas}{samples and features data frame.}

\item{dt.Atlas.clinical}{clinical data as data frame.}

\item{list.clinical.projection}{(list) list of clinical data to project on
the map.}

\item{path.maps}{path to the SOM map to loaded to project data.}

\item{model.name}{name of the current MOFA model to use.}

\item{nb.SOM.maps}{(int) number of SOM map generated.}

\item{nb.samples}{(description)int) number of sample.}
}
\description{
This function project clinical data (as data frame) in SOM map
loaded.
}
