% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProjectionClusterAnalysis.R
\name{ProjectionClusterAnalysis}
\alias{ProjectionClusterAnalysis}
\title{Analysis of cluster.}
\usage{
ProjectionClusterAnalysis(
  nb.cluster,
  path.maps,
  model.name,
  nb.SOM.maps,
  dt.Atlas,
  path.best = "4_Best_maps_consensus",
  list.view = NULL,
  threshold = 0.05,
  p.adj.threshold = T,
  log2 = F,
  all.feature.test = T,
  hierarchical.test = T,
  plot.box.plot = T,
  color.cluster = NULL
)
}
\arguments{
\item{nb.cluster}{(int) number of cluster.}

\item{path.maps}{path to the maps directory.}

\item{model.name}{(str) name of the current MOFA model to analyse.}

\item{nb.SOM.maps}{(int) number of generated SOM map.}

\item{dt.Atlas}{Atlas data frame data.}

\item{path.best}{(string) path to the folder to save best maps.}

\item{list.view}{pass a list of views to analyse. By default is "NULL"' (do 
analyse for all views).}

\item{threshold}{(float) threshold for select significant features regarding
p value.}

\item{p.adj.threshold}{(bool) to use adjusted p value.}

\item{log2}{(bool) to add log fold change value.}

\item{all.feature.test}{(bool) if we compute all features test for cluster
analyse.}

\item{hierarchical.test}{(bool) if we compute hierarchical test for cluster
analysis.}

\item{plot.box.plot}{(bool) if plot box plot figures as pdf.}

\item{color.cluster}{color palette for each cluster.}
}
\description{
Cluster analysis for current MOFA model.
}
