% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProjectionClusterAnalysisBoxPlot.R
\name{ProjectionClusterAnalysisBoxPlot}
\alias{ProjectionClusterAnalysisBoxPlot}
\title{Function to define box plot parameters and generate figure.}
\usage{
ProjectionClusterAnalysisBoxPlot(
  df.data,
  df.result,
  nb.cluster,
  view,
  color.cluster,
  path.result,
  p.adj.threshold = T,
  threshold = 0.05,
  log2 = F
)
}
\arguments{
\item{df.data}{Atlas data frame data.}

\item{df.result}{(dataframe) data frame with results for box plot values.}

\item{nb.cluster}{(int) nth cluster to analyse.}

\item{view}{(str) current view of data.}

\item{color.cluster}{(str) color.cluster for studied cluster.}

\item{path.result}{path to save result.}

\item{p.adj.threshold}{(bool) if we use adjust p value.}

\item{threshold}{p value threshold.}

\item{log2}{(bool) if compute folder log2 change.}
}
\description{
We initialize parameters for box l=plot figure and then we run
visualize function to generate box plot.
}
