% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProjectionClusterAnalysisFeaturesTest.R
\name{ProjectionClusterAnalysisFeaturesTest}
\alias{ProjectionClusterAnalysisFeaturesTest}
\title{Compute cluster features correlation test.}
\usage{
ProjectionClusterAnalysisFeaturesTest(
  data,
  clusters,
  path.result,
  threshold = 0.05,
  view = NULL,
  log2 = T,
  p.adj.threshold = T,
  nb.cluster = 5,
  all.feature.test = T,
  hierarchical.test = T,
  plot.box.plot = T,
  color.cluster = NULL
)
}
\arguments{
\item{data}{Atlas data frame data.}

\item{clusters}{clusters consensus as csv.}

\item{path.result}{path to the folder where the outputs are saved.}

\item{threshold}{p value threshold.}

\item{view}{pass a list of views to analyse. By default is "NULL"' (do 
analyse for all views).}

\item{log2}{(bool) if compute folder log2 change.}

\item{p.adj.threshold}{(bool) if we use adjust p value.}

\item{nb.cluster}{(int) number of cluster.}

\item{all.feature.test}{(bool) if we compute all features test for cluster
analyse.}

\item{hierarchical.test}{(bool) if we compute hierarchical test for cluster
analysis.}

\item{plot.box.plot}{(bool) if plot box plot figures as pdf.}

\item{color.cluster}{color palette for each cluster.}
}
\description{
We compute test about correlation between features and cluster.
}
