% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProjectionConsensusMap.R
\name{ProjectionConsensusMap}
\alias{ProjectionConsensusMap}
\title{Create SOM maps with feature values}
\usage{
ProjectionConsensusMap(
  parallel.run,
  nb.cluster,
  path.maps,
  model.name,
  path.model.factor,
  path.best = "4_Best_maps_consensus",
  keep.groups = NULL,
  groups = NULL,
  regularized = T,
  show.samples = F,
  all.features = T,
  add.intercept = T,
  hierarchical.clustering = F,
  kmeans.clustering = T
)
}
\arguments{
\item{parallel.run}{config to run script in parallel}

\item{nb.cluster}{(int) number of cluster}

\item{path.maps}{maps path on which to project feature values}

\item{model.name}{name of the current analysed model}

\item{path.model.factor}{path to the model for a specific factor number.}

\item{path.best}{(string) path to the folder to save best maps.}

\item{keep.groups}{(list) list of dat agroups to keep dor analysis.}

\item{groups}{(list) list of group to which we want the data to belong}

\item{regularized}{(bool) if we regularized data.}

\item{show.samples}{(bool) if we plot projection sample in the figure.}

\item{all.features}{(bool) if we do analysis for all features}

\item{add.intercept}{(bool) if add intercept (default=T).}

\item{hierarchical.clustering}{(bool) if use hierarchical clustering 
algorithm (default=F)}

\item{kmeans.clustering}{(bool) if use hierarchical clustering 
algorithm (default=T)}
}
\description{
We create SOM map with feature values.
}
