% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProjectionFeature.R
\name{ProjectionFeature}
\alias{ProjectionFeature}
\title{Feature projection on existing SOM maps}
\usage{
ProjectionFeature(model, data.som, path.save, dt.feature, groups, feature)
}
\arguments{
\item{model}{MOFA model trained.}

\item{data.som}{SOM map codes matrix.}

\item{path.save}{path to save plot.}

\item{dt.feature}{data frame with feature values.}

\item{groups}{specify which groups should be kept in the model.}

\item{feature}{feature name to plot.}
}
\description{
We project feature values on existing map.
}
