% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProjectionGetSOMNodes.R
\name{ProjectionGetSOMNodes}
\alias{ProjectionGetSOMNodes}
\title{Associate nearest SOM neuron for each Z rows.}
\usage{
ProjectionGetSOMNodes(dt.Z, data.som)
}
\arguments{
\item{dt.Z}{Z data frame to find nearest neuron.}

\item{data.som}{SOM map}
}
\description{
Compute the nearest neuron unit for each Z rows on the SOM map.
}
