% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProjectionMapToView.R
\name{ProjectionMapToView}
\alias{ProjectionMapToView}
\title{return predicted data from the MOFA model trained}
\usage{
ProjectionMapToView(
  data.som.codes,
  model,
  regularized = F,
  views = "all",
  groups = "all",
  factors = "all",
  add.intercept = T
)
}
\arguments{
\item{data.som.codes}{matrix of codes for each current SOM map.}

\item{model}{current MOFA model trained.}

\item{regularized}{(bool) if Z matrix data are regularized or not.}

\item{views}{(list) list of views to which we want the data to belong}

\item{groups}{(list) list of group to which we want the data to belong}

\item{factors}{(list) list of factors to which we want the data to belong}

\item{add.intercept}{(bool) if add intercept}
}
\value{
predicted.data
}
\description{
return predicted data from the current MOFA model trained.
}
