% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProjectionNewData.R
\name{ProjectionNewData}
\alias{ProjectionNewData}
\title{Function to compute coefficient for new data projection.}
\usage{
ProjectionNewData(data.new, model, path.model, coef.optimized = NULL)
}
\arguments{
\item{data.new}{data frame with new data to be projected.}

\item{model}{the trained MOFA model to be used to project the new data.}

\item{path.model}{path to the model to save coef.}

\item{coef.optimized}{coeff to weight Z matrix.}
}
\value{
Z.new.projection
}
\description{
This function optimize coefficients to align new Z projection 
data with original Z reference projection.
To optimize coefficient, we have three normalization mode for coefficient:
With mode 1, we normalize coefficient with sum(coef**2). With mode 2, we 
normalize coefficient with sum(coef[1:length(Zproj)]), we don't take into
count intercept. And with mode 3, we normalize coefficient with 
sum(coef[1:length(Zproj)]**2), we don't take into count intercept.
If we have already compute this coefficients, this function can return 
new Z projection data frame computed from this coefficients.
}
