% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProjectionSOMClusterOverLighting.R
\name{ProjectionSOMClusterOverLighting}
\alias{ProjectionSOMClusterOverLighting}
\title{Function to create a figure with the cluster zones highlighted on the 
SOM map.}
\usage{
ProjectionSOMClusterOverLighting(
  nb.cluster,
  path.maps,
  model.name,
  nb.samples,
  main.cluster.color = "#66CC00",
  background.cluster.color = "#E0E0E0"
)
}
\arguments{
\item{nb.cluster}{(int) numerous of cluster to highlight.}

\item{path.maps}{path to the map folder.}

\item{model.name}{name og the current MOFA model.}

\item{nb.samples}{(description)int) number of sample.}

\item{main.cluster.color}{(str) Hexadecimal code to define color for the main
cluster to highlight.}

\item{background.cluster.color}{(str) Hexadecimal code to define color for 
others background clusters.}
}
\description{
This function generates a pdf with the SOM map of the current
MOFa model. Nodes in the map for a chosen cluster are highlighted in green 
by default, and other nodes in other clusters are highlighted in gray by 
default.
Different colors can be selected with the main.cluster.color and 
background.cluster.color parameters.
This function generates a pdf image for each map in the 
'4_Best_maps_consensus' folder.
}
