% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProjectionSOMData.R
\name{ProjectionSOMData}
\alias{ProjectionSOMData}
\title{Project Clinical data passed in argument in SOM map.}
\usage{
ProjectionSOMData(
  projected,
  map,
  path.directory,
  dt.Atlas.clinical,
  list.clinical.projection,
  groups,
  nb.samples
)
}
\arguments{
\item{projected}{data frame of the data to project on the map.}

\item{map}{current SOM map to use.}

\item{path.directory}{path to save figure projection.}

\item{dt.Atlas.clinical}{clinical data as data frame.}

\item{list.clinical.projection}{(list) list of clinical data to project on
the map.}

\item{groups}{data frame with cluster value of each nodes of the SOM map.}

\item{nb.samples}{(int) number of samples.}
}
\description{
This function project clinical data (as data frame) in SOM map
loaded.
}
