% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOMBestMapAnalysis.R
\name{SOMBestMapAnalysis}
\alias{SOMBestMapAnalysis}
\title{Additional analyse of best maps selected based on scores.}
\usage{
SOMBestMapAnalysis(
  nb.cluster,
  path.maps,
  model.name,
  nb.SOM.maps,
  nb.best.maps.to.keep = 3,
  best.map.type = "ARI",
  verbose = F,
  clustering.method = "hierarchical_clustering"
)
}
\arguments{
\item{nb.cluster}{(int) number of cluster to test.}

\item{path.maps}{path to maps folder.}

\item{model.name}{name of the current model to analysis.}

\item{nb.SOM.maps}{(int) number of maps generated.}

\item{nb.best.maps.to.keep}{(int) number of the first best maps to keep 
(for each score).}

\item{best.map.type}{(str) type of score to use for best maps selection.}

\item{verbose}{(bool) if function print messages or not.}

\item{clustering.method}{(str) clustering method used.}
}
\description{
We add more specific analysis for selected maps based on 
predefined scores.
In a first step we select the n best maps according to the scores we have 
calculated (diagonal score and score). 
Then, for each of these selected maps, we copy the original folder of the 
selected map to a new location (in order to facilitate its access later). 
We then merge the values of the consensus clustering with the values of the 
original clustering, keeping the consensus values.
We obtain a new clustering based on the consensus clustering of the current 
model with the current number of clusters.
We then create the new groups according to this new clustering to create the 
new figures.
}
