% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOMComputeFitToConsensus.R
\name{SOMComputeFitToConsensus}
\alias{SOMComputeFitToConsensus}
\title{Check of the cluster entity on the map}
\usage{
SOMComputeFitToConsensus(
  new.groups,
  nb.clusters,
  path.map.directory,
  cluster.consensus,
  cluster.new,
  model.name,
  matrix.som.pts,
  nth.map,
  add.quantization.score = T,
  data.som = NULL
)
}
\arguments{
\item{new.groups}{clustering.}

\item{nb.clusters}{(int) current number of cluster.}

\item{path.map.directory}{path to save consensus results.}

\item{cluster.consensus}{data frame of consensus clustering.}

\item{cluster.new}{data frame with new clustering obtained with new.groups.}

\item{model.name}{name of current model.}

\item{matrix.som.pts}{nodes SOM coordinates as matrix.}

\item{nth.map}{(int) current map number.}

\item{add.quantization.score}{(bool) if we add quantization score to measure
best som map.}

\item{data.som}{matrix with SOM map factors values.}
}
\description{
check that the clusters are not split into two or more parts.
This is checked in order to respect the intuition that the samples physically
close on the map must be biologically similar and therefore they must belong
to the same cluster.
We define also the concordance to consensus score to measure how close 
the map is to the consensus map.
}
