% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOMDefineBEstModelConsensus.R
\name{SOMDefineBEstModelConsensus}
\alias{SOMDefineBEstModelConsensus}
\title{Define best model consensus}
\usage{
SOMDefineBEstModelConsensus(
  path.maps,
  model.name,
  nb.SOM.maps,
  list.nb.clusters,
  use.quantization.error = T,
  concordance.precision = 2,
  file.name = "Best_model.consensus.csv",
  clustering.method = "hierarchical_clustering"
)
}
\arguments{
\item{path.maps}{path to the maps folder}

\item{model.name}{name of the current model}

\item{nb.SOM.maps}{number of generated maps
#' @param use.quantization.error (bool) if TRUE, we take into count the 
quantization error of the map. Quantization error measures how well the SOM 
represents the input data.}

\item{concordance.precision}{is the precision to sort best map for 
concordance score, i.e. we round all concordance values with this precision 
(To round a value to n digits after the decimal point, by default is 2) and
for all same score we sort regarding the quantization error.}

\item{file.name}{(str) name of the data frame to save with fit scores.}

\item{clustering.method}{(str) clustering method used.}
}
\description{
This function allow to get the n first best map for the best 
model according concordance score.
We copy this best maps in folder best maps consensus.
}
