% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOMDefineNewConsensusBoundaries.R
\name{SOMDefineNewConsensusBoundaries}
\alias{SOMDefineNewConsensusBoundaries}
\title{Define new clustering boundaries based on consensus clustering}
\usage{
SOMDefineNewConsensusBoundaries(groups, nb.clusters, Clusters, matrix.som.pts)
}
\arguments{
\item{groups}{data frame with cluster for SOM nodes}

\item{nb.clusters}{(int) current number of cluster}

\item{Clusters}{data frame with clustering consensus information}

\item{matrix.som.pts}{matrix SOM nodes}
}
\value{
rank.matrix.cluster
}
\description{
Use a function like "heat function" to establish the new 
clustering boundaries from the consensus clustering.
In order to simulate a kind of heat dispersion with the already known 
clusters, we propagate the clustering from the nodes we know to the nodes 
we don't know with the while loop (see the function below).
The variable "nb.indeter" allows us to know the count of the nodes that don't
have an associated cluster yet and so we continue the dispersal as long as
all the nodes are not filled in.
}
