% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOMGetConsensusModel.R
\name{SOMGetConsensusModel}
\alias{SOMGetConsensusModel}
\title{Create data frame with all concordance to consensus scores}
\usage{
SOMGetConsensusModel(
  path.maps,
  model.name,
  nb.cluster,
  nb.SOM.maps,
  clustering.method = "hierarchical_clustering"
)
}
\arguments{
\item{path.maps}{path to maps folder}

\item{model.name}{name of current model}

\item{nb.cluster}{(int) current number of cluster}

\item{nb.SOM.maps}{number of generated maps}

\item{clustering.method}{(str) clustering method used.}
}
\value{
model.consensus
}
\description{
This function stack all concordance to consensus scores of all
nth.map map for one model and return it.
}
