% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOMGetNewConsensusClusteringElements.R
\name{SOMGetNewConsensusClusteringElements}
\alias{SOMGetNewConsensusClusteringElements}
\title{Define new group values}
\usage{
SOMGetNewConsensusClusteringElements(
  groups,
  nb.clusters,
  clusters,
  clusters.original,
  matrix.som.pts,
  path.map.directory
)
}
\arguments{
\item{groups}{data frame with cluster for SOM nodes}

\item{nb.clusters}{(int) current number of cluster}

\item{clusters}{data frame with clustering consensus information}

\item{clusters.original}{original cluster}

\item{matrix.som.pts}{matrix SOM nodes}

\item{path.map.directory}{path to save consensus results}
}
\value{
new.groups
}
\description{
From rank matrix result, in this function we define the new
groups values for SOM nodes.
}
