% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOMGetNodeNeighbors.R
\name{SOMGetNodeNeighbors}
\alias{SOMGetNodeNeighbors}
\alias{.SOMGetNodeNeighbors}
\title{Find out node's neighbors}
\usage{
.SOMGetNodeNeighbors(
  current_node_number,
  matrix_som_pts,
  toroidal = F,
  eps = 0.8660254
)
}
\arguments{
\item{current_node_number}{(int) the numerous of node to find neighbors.}

\item{matrix_som_pts}{matrix SOM nodes.}

\item{toroidal}{(bool) If toroidal (i.e. the ends of the map meet and the 
values are therefore continuous.).}

\item{eps}{(float) physical length of one node on y axis.}
}
\value{
(list) neighbors nodes
}
\description{
Return the numerous of node which are neighbor of current node 
passed in parameter.
}
