% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOMJaccardAnalysis.R
\name{SOMJaccardAnalysis}
\alias{SOMJaccardAnalysis}
\title{Plot and save Jaccard analysis to compare models}
\usage{
SOMJaccardAnalysis(path.result, jaccard.index.total, list.nb.clusters)
}
\arguments{
\item{path.result}{path to save jaccard index data frame.}

\item{jaccard.index.total}{(csv) data frame with different jaccard indexes.}

\item{list.nb.clusters}{(list) list of cluster number to make analysis}
}
\description{
This function generate different plot and data frame to compare
models based on Jaccard scores analysis.
}
