% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOMMapsToKeep.R
\name{SOMMapsToKeep}
\alias{SOMMapsToKeep}
\title{Load and keep best maps based on early computed scores}
\usage{
SOMMapsToKeep(
  path.result,
  nb.SOM.maps,
  nb.best.maps.to.keep,
  keep.baseline.maps = F,
  keep.ARI.maps = T
)
}
\arguments{
\item{path.result}{path to result folder}

\item{nb.SOM.maps}{(int) number of generated maps}

\item{nb.best.maps.to.keep}{(int) number of the first best maps to keep}

\item{keep.baseline.maps}{(bool) if we return best map according baseline
score.}

\item{keep.ARI.maps}{(bool) if we return best map according ARI score.}
}
\value{
keep.maps
}
\description{
This function return the nb.best.maps.to.keep first maps 
with the best scores.
We keep n_score * nb.best.maps.to.keep maps (nb.best.maps.to.keep maps for 
"diag" score and nb.best.maps.to.keep maps for "score" score).
}
