% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOMMergeOriginalConsensusClustering.R
\name{SOMMergeOriginalConsensusClustering}
\alias{SOMMergeOriginalConsensusClustering}
\title{Merge original and consensus clustering to return new consensus group
and clustering.}
\usage{
SOMMergeOriginalConsensusClustering(
  path.map.original,
  path.map.consensus,
  nb.SOM.maps,
  clustering.method
)
}
\arguments{
\item{path.map.original}{path to the original folder.}

\item{path.map.consensus}{path to the consensus folder.}

\item{nb.SOM.maps}{number of generated maps to compute consensus clustering.}

\item{clustering.method}{(str) clustering methode used.}

\item{return}{a list of cluster matrix (merged clustering, original 
clustering and consensus clustering).}
}
\description{
This function load original and clustering consensus matrix
of a specific map. Then merge the both matrix and return a list of all 
clustering.
}
