% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOMSelectAndCopyBestMaps.R
\name{SOMSelectAndCopyBestMaps}
\alias{SOMSelectAndCopyBestMaps}
\title{Copy n first best map for current model based on concordance to 
consensus score}
\usage{
SOMSelectAndCopyBestMaps(
  model.consensus,
  path.result,
  path.maps,
  path.best.maps,
  n.first.map = 10,
  use.quantization.error = T,
  concordance.precision = 2,
  file.name = "best_model_consensus.csv",
  clustering.method = "hierarchical_clustering"
)
}
\arguments{
\item{model.consensus}{consensus clustering for current model}

\item{path.result}{path to the result folder}

\item{path.maps}{the path to maps}

\item{path.best.maps}{path to the folder with best map saved}

\item{n.first.map}{(int) the n first map to save in the target result folder}

\item{use.quantization.error}{(bool) if TRUE, we take into count the 
quantization error of the map. Quantization error measures how well the SOM 
represents the input data.}

\item{concordance.precision}{is the precision to sort best map for 
concordance score, i.e. we round all concordance values with this precision 
(To round a value to n digits after the decimal point, by default is 2) and
for all same score we sort regarding the quantization error.}

\item{file.name}{(str) name of the data frame to save with fit scores.}

\item{clustering.method}{(str) clustering method used.}
}
\description{
We order all best map based on concordance score and we copy
the "n.first.map" in target folder.
In this function we also save a data frame with best mode consensus.
If we use quantization error, we sort best map first by 'over_cluster' value,
then by concordance to consensus score and then by quantization error.
In this case (use.quantization.error=T) we use precision digit to sort value.
}
