% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOMTransformToConsensusMap.R
\name{SOMTransformToConsensusMap}
\alias{SOMTransformToConsensusMap}
\title{Determine and update best clustering for best map.}
\usage{
SOMTransformToConsensusMap(
  parallel.run,
  model.name,
  path.maps,
  nb.SOM.maps,
  list.nb.clusters,
  verbose = F,
  clustering.method = "hierarchical_clustering"
)
}
\arguments{
\item{parallel.run}{computing parameter for parallel running.}

\item{model.name}{current model with current list.nb.clusters and current
nth.map.}

\item{path.maps}{path to the maps folder.}

\item{nb.SOM.maps}{(int) number of generated maps.}

\item{list.nb.clusters}{(list) list number of cluster to run.}

\item{verbose}{(bool) if function print messages or not.}

\item{clustering.method}{(str) clustering method used.}
}
\description{
In this function we transform all original maps with new
consensus clustering values and we compute new consensus boundaries on 
the original map.
Then, we compute scores regarding the fit between the new consensus 
clustering of the samples and clustering of the node (determined by the new
consensus boundaries).
}
