% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValidatePath.R
\name{ValidatePath}
\alias{ValidatePath}
\title{Function to test path to save file}
\usage{
ValidatePath(path, nb.char.max = 182)
}
\arguments{
\item{path}{The path to the file to be saved.}
}
\value{
A boolean value indicating whether the path is valid.
}
\description{
Function to test path to save file
}
